feat: explicit video refresh
Instead of regularly sending headers and keyframes every few seconds,
the idea here is to send an endless stream of P frames, each referencing
the previous one. This saves some bandwidth, but means that a client
which "falls off" the chain has no way of (automatically) recovering.

To fix that, the client needs an explicit way to request a
keyframe/header packet, which is added here as an attachment message.
This is much more robust for the client than just waiting for a refresh
to come along, and results in shorter pauses when an important frame is
dropped.

Because we use hierarchical coding (for h265), we take care to indicate
if a frame is actually important, so that clients can avoid requesting
refreshes when the stream is already recoverable.

+++ mm-server/src/compositor/video/vulkan_encode/rate_control.rs
@@ -67,7 +67,7 @@ impl LayeredVbr {
             return self.base;
         }
 
-        let bitrate_denominator = 2_u64.pow(layer as u32 + 1);
+        let bitrate_denominator = 2_u64.pow(layer + 1);
         let max_qp = layer_qp(self.base.max_qp, layer).clamp(self.base.min_qp, self.base.max_qp);
 
         VbrSettings {
